Move existing table elements to controls/tables
