+++ src/components/controls/tables/FancyTable/package.json
